# tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
      "ArtifactDetail": {
        "description": "A single artifact: its projection, its full catalog record, and what is attested.\n\n``attestations`` names the attestation *types present in the registry* \u2014 it is emphatically not\na verification verdict, and the front end is required to say so in those words (ui.md \u00a77).\nEmpty when the deployment has no registry to ask.",
        "properties": {
          "artifact_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Kind"
          },
          "attestations": {
            "items": {
              "type": "string"
            },
            "title": "Attestations",
            "type": "array"
          },
          "deprecated": {
            "default": false,
            "title": "Deprecated",
            "type": "boolean"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "record": {
            "additionalProperties": true,
            "title": "Record",
            "type": "object"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "yanked": {
            "default": false,
            "title": "Yanked",
            "type": "boolean"
          }
        },
        "required": [
          "reference",
          "digest",
          "name",
          "version",
          "score",
          "record"
        ],
        "title": "ArtifactDetail",
        "type": "object"
      },
...
      "BenchJobRecord": {
        "additionalProperties": false,
        "description": "Bench's evaluation job record, named for the document rather than for the module.\n\nBench and Studio each own a ``JobRecord``, and they are genuinely different things \u2014 one tracks\nan evaluation, the other a design study. Mounted in one process they collide in the OpenAPI\ndocument, and FastAPI disambiguates by prefixing the module path, which produces\n``astro_mine__bench__leaderboard___jobs__JobRecord`` and a generated client method returning a\ntype nobody can type.\n\nThe collision is an artifact of *composition* \u2014 it exists only because this distribution serves\nboth surfaces from one document \u2014 so it is resolved here rather than by renaming a platform type\nthat is unambiguous on its own. This subclass adds nothing: every field is inherited, so the\ndocument and the platform model cannot drift.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "status": {
            "$ref": "#/components/schemas/SubmissionStatus"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "BenchJobRecord",
        "type": "object"
      },
...
      "DownloadBody": {
        "description": "What a requester presents at the gated download boundary.",
        "properties": {
          "allowed_licenses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Licenses"
          },
          "grants": {
            "items": {
              "type": "string"
            },
            "title": "Grants",
            "type": "array"
          },
          "require_verified": {
            "default": false,
            "title": "Require Verified",
            "type": "boolean"
          }
        },
        "title": "DownloadBody",
        "type": "object"
      },
      "DownloadGrant": {
        "description": "Permission to materialize an artifact, with the policy that granted it.\n\nThe policy version travels with the grant so a consumer can record which rules let it in.",
        "properties": {
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "policy_engine": {
            "title": "Policy Engine",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          }
        },
        "required": [
          "digest",
          "policy_version",
          "policy_engine"
        ],
        "title": "DownloadGrant",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HubSubmissionRequest": {
        "additionalProperties": false,
        "description": "A community submission referenced **only by Hub digest** (RM-P1-BENCH-10; bench.md \u00a76).\n\n``hub_ref`` is a Hub reference \u2014 a ``name:version`` tag or a ``sha256:`` image-manifest digest \u2014\nthat Bench resolves from Hub, **verifies fail-closed** (content address, then cosign signature +\nSLSA provenance + SBOM \u2014 bench#29), whose Core plugin manifest is validated against the scenario\ninterface, and which is then run under submit-policy-we-run **inside a sandbox** (bench#30). No\npolicy bytes are uploaded: the artifact is authenticated by content hash and by signature.\n\n``method``/``author`` are display metadata only. There is deliberately **no ``identity``\nfield**: the submitter's identity comes from the verified OIDC bearer token, and rate limits,\nquotas, job tickets, and audit records are all keyed on that (bench#29). The pre-bench#29 model\ncarried a client-supplied ``identity`` that keyed the rate limiter \u2014 so a submitter could reset\ntheir own quota by editing a JSON field.",
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "hub_ref": {
            "minLength": 1,
            "title": "Hub Ref",
            "type": "string"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "scenario_id": {
            "minLength": 1,
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "hub_ref"
        ],
        "title": "HubSubmissionRequest",
        "type": "object"
      },
...
      "PublishBody": {
        "description": "Index a Core plugin manifest for an already-stored artifact.",
        "properties": {
          "digest": {
            "minLength": 1,
            "title": "Digest",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "namespace": {
            "default": "open",
            "title": "Namespace",
            "type": "string"
          },
          "publisher": {
            "minLength": 1,
            "title": "Publisher",
            "type": "string"
          }
        },
        "required": [
          "manifest",
          "digest",
          "publisher"
        ],
        "title": "PublishBody",
        "type": "object"
      },
...
      "ResolveBody": {
        "description": "A resolution constraint set.",
        "properties": {
          "capability_tags": {
            "items": {
              "type": "string"
            },
            "title": "Capability Tags",
            "type": "array"
          },
          "interfaces": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interfaces"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version_spec": {
            "default": "",
            "title": "Version Spec",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ResolveBody",
        "type": "object"
      },
...
      "SearchHit": {
        "description": "One catalog entry as the API projects it \u2014 the shape ``_hit`` builds.\n\nThis is a *projection* of :class:`~astro_mine.hub.index.CatalogEntry`, not a copy of it: the\nroute already chose these twelve fields, and declaring them makes the choice legible to a\ngenerated client instead of leaving it as an untyped object the client types as ``unknown``.\nThe catalog record itself is not here \u2014 it is large, and only the detail route returns it.",
        "properties": {
          "artifact_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Kind"
          },
          "deprecated": {
            "default": false,
            "title": "Deprecated",
            "type": "boolean"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "license": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "yanked": {
            "default": false,
            "title": "Yanked",
            "type": "boolean"
          }
        },
        "required": [
          "reference",
          "digest",
          "name",
          "version",
          "score"
        ],
        "title": "SearchHit",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "The Astro-Mine REST tier: the Hub registry API, the Studio API, Cloud's submission edge and the Bench leaderboard, behind one deployable.",
    "title": "Astro-Mine-API",
    "version": "<version>"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bench/submissions/hub": {
      "post": {
        "description": "Submit a community artifact by Hub digest; verified (cosign/SLSA/SBOM) then sandboxed.",
        "operationId": "bench_submit_hub",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HubSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchJobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Hub",
        "tags": [
          "submissions"
        ]
      }
    },
...
    "/hub/artifacts/{name}/{version}": {
      "get": {
        "operationId": "hub_get_artifact",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifact"
      }
    },
    "/hub/artifacts/{name}/{version}/download": {
      "post": {
        "operationId": "hub_download",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadGrant"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download"
      }
    },
...
    "/hub/publish": {
      "post": {
        "description": "Index an already-stored artifact \u2014 after proving the caller's claims about it.\n\nEvery field in the body is a *claim*: the digest, the manifest, and the namespace. This\nendpoint used to take all three on the caller's word, which let a request forge content\nprovenance rather than merely omit it. Admission now re-derives each from the registry\n(hub.md \u00a72.3), and a caller-asserted trust tier above ``open`` is refused outright \u2014\npromotion is a curated, audited action (hub.md \u00a79), never a field in a publish request.",
        "operationId": "hub_publish",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchHit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish"
      }
    },
    "/hub/resolve": {
      "post": {
        "operationId": "hub_resolve",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Do Resolve"
      }
    },
    "/hub/search": {
      "get": {
        "operationId": "hub_search",
        "parameters": [
          {
            "in": "query",
            "name": "text",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Text"
            }
          },
          {
            "in": "query",
            "name": "semantic",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semantic"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "artifact_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Kind"
            }
          },
          {
            "in": "query",
            "name": "license",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "License"
            }
          },
          {
            "in": "query",
            "name": "namespace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Namespace"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchHit"
                  },
                  "title": "Response Hub Search",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Do Search"
      }
    },
...
  }
}